+++ packages/transaction-pay-controller/src/utils/quotes.test.ts
@@ -1,4 +1,7 @@
-import type { TransactionMeta } from '@metamask/transaction-controller';
+import {
+  TransactionStatus,
+  type TransactionMeta,
+} from '@metamask/transaction-controller';
 import type { BatchTransaction } from '@metamask/transaction-controller';
 import type { Hex, Json } from '@metamask/utils';
 import { cloneDeep } from 'lodash';
@@ -48,6 +51,7 @@ const TRANSACTION_DATA_MOCK: TransactionData = {
 
 const TRANSACTION_META_MOCK = {
   id: TRANSACTION_ID_MOCK,
+  status: TransactionStatus.unapproved,
   txParams: { from: '0xabc' as Hex },
 } as TransactionMeta;
 
@@ -94,9 +98,10 @@ describe('Quotes Utils', () => {
    * Run the updateQuotes function.
    *
    * @param params - Partial params to override the defaults.
+   * @returns Return value from updateQuotes.
    */
   async function run(params?: Partial<UpdateQuotesRequest>) {
-    await updateQuotes({
+    return await updateQuotes({
       messenger,
       transactionData: cloneDeep(TRANSACTION_DATA_MOCK),
       transactionId: TRANSACTION_ID_MOCK,
@@ -299,6 +304,18 @@ describe('Quotes Utils', () => {
           totalFiat: TOTALS_MOCK.total.usd,
         },
       });
+    });
+
+    it('does nothing if transaction is not unapproved', async () => {
+      getTransactionMock.mockReturnValue({
+        ...TRANSACTION_META_MOCK,
+        status: TransactionStatus.confirmed,
+      });
+
+      const result = await run();
+
+      expect(updateTransactionDataMock).not.toHaveBeenCalled();
+      expect(result).toBe(false);
     });
   });